+++ src/oklch/colors.py
@@ -37,8 +37,10 @@ class Color:
     # These return a dummy color just so to eliminate an annoying warning. 
     def to_RGB(self): 
         return RGB(0, 0, 0)
-    def to_Hex(self):
-        return Hex('#000000')
+    def to_HEX(self):
+        return HEX('#000000')
+    def to_OKLAB(self):
+        return OKLAB(0, 0, 0)
     def to_OKLCH(self):
         return OKLCH(0, 0, 0)
 
@@ -47,22 +49,36 @@ class Color:
 
     def __str__(self): return ""
 
-    # Subtraction yields the euclidean distance between the two
+    # Checks that arg is color
+    @staticmethod
+    def __is_color(arg):
+        if not isinstance(arg, Color):
+            raise ValueError(f"Expected color, received '{type(arg)}'!")
+    # Addition gives the midpoint of the two colors in OKLAB space
+    def __add__(self, other):
+        self.__is_color(other)
+
+        return self.to_OKLAB() + other.to_OKLAB()
+    # Negation gives the complement in OKLAB space
+    def __neg__(self):
+        return self.to_OKLAB().__neg__()
+    # Subtraction gives the midpoint of self and complement of other
     def __sub__(self, other):
-        if not isinstance(other, Color):
-            msg = "Expected colors, received '" + str(type(self)) \
-                    + "' and '" + str(type(other)) + "'!"
-            raise ValueError(msg)
+        self.__is_color(other)
+
+        return self.to_OKLAB() - other.to_OKLAB()
+
+    # Pipe operator yields the euclidean distance between two colors
+    def __or__(self, other):
+        self.__is_color(other)
 
         # Convert both colors to oklab:
-        self = self.to_OKLCH()
-        s = self._rect()
-        other = other.to_OKLCH()
-        o = other._rect()
+        self = self.to_OKLAB()
+        other = other.to_OKLAB()
 
-        return math.pow((self.l - other.l)**2 + \
-                            (s[0] - o[0])**2 + \
-                            (s[1] - o[1])**2,
+        return math.pow((self.l - other.l) ** 2 \
+                            + (self.a - other.a) ** 2 \
+                            + (self.b - other.b) ** 2,
                         0.5)
 
     ColorDict = { \
@@ -210,12 +226,10 @@ class Color:
 
     @staticmethod
     def get_web_color(color_name):
-        return Hex(Color.ColorDict[color_name])
-
+        return HEX(Color.ColorDict[color_name])
     @staticmethod
     def get_random_web_color():
         return Color.get_web_color(choice(list(Color.ColorDict.keys())))
-
     @staticmethod
     def get_nearest_web_color(color):
         if not isinstance(color, Color):
@@ -235,16 +249,15 @@ class Color:
         # Build minheap of the web colors by distance
         heap = []
         for k, v in Color.ColorDict.items():
-            c = Hex(v).to_OKLCH()
-            heapq.heappush(heap, Node(color - c, k, c))
+            c = HEX(v).to_OKLCH()
+            heapq.heappush(heap, Node(color | c, k, c))
 
         ret = heapq.heappop(heap)
         return (ret.color_name, ret.color)
 
 ###############################################################################
 #
-# Original license for RGB.to_OKLCH(), OKLCH._f(), OKLCH._f_inv(),
-#   OKLCH.to_RGB():
+# Original license for RGB.to_OKLAB(), RGB._f(), RGB._f_inv(), OKLAB.to_RGB():
 #
 #   Copyright (c) 2021 Björn Ottosson
 #   
@@ -278,84 +291,25 @@ class RGB(Color):
     def __str__(self):
         return "rgb({}, {}, {})".format(self.r, self.g, self.b)
 
+    # Return type for addition and subtraction is type of first operand
+    def __add__(self, other):
+        return super().__add__(other).to_RGB()
+    def __neg__(self):
+        return super().__neg__().to_RGB()
     def __sub__(self, other):
-        return super().__sub__(other)
+        return super().__sub__(other).to_RGB()
 
+    def __or__(self, other):
+        return super().__or__(other)
+
+    # Type Conversions
     def to_RGB(self):
         return self
-
-    def to_Hex(self):
-        return Hex("#{:0>2}{:0>2}{:0>2}".format(
+    def to_HEX(self):
+        return HEX("#{:0>2}{:0>2}{:0>2}".format(
                 _hex(self.r),
                 _hex(self.g),
                 _hex(self.b)))
-
-    def to_OKLCH(self):
-        l = 0.4122214708 * OKLCH._f_inv(self.r/255) + \
-            0.5363325363 * OKLCH._f_inv(self.g/255) + \
-            0.0514459929 * OKLCH._f_inv(self.b/255)
-        m = 0.2119034982 * OKLCH._f_inv(self.r/255) + \
-            0.6806995451 * OKLCH._f_inv(self.g/255) + \
-            0.1073969566 * OKLCH._f_inv(self.b/255)
-        s = 0.0883024619 * OKLCH._f_inv(self.r/255) + \
-            0.2817188376 * OKLCH._f_inv(self.g/255) + \
-            0.6299787005 * OKLCH._f_inv(self.b/255)
-
-        l_ = math.pow(l, 1/3)
-        m_ = math.pow(m, 1/3)
-        s_ = math.pow(s, 1/3)
-
-        return OKLCH._polarize(
-            0.2104542553*l_ + 0.7936177850*m_ - 0.0040720468*s_,
-            1.9779984951*l_ - 2.4285922050*m_ + 0.4505937099*s_,
-            0.0259040371*l_ + 0.7827717662*m_ - 0.8086757660*s_)
-
-    # Check whether the color is in-gamut
-    def is_in_gamut(self):
-        return max(self.r, self.g, self.b) <= 255 and \
-                min(self.r, self.g, self.b) >= 0
-
-# RGB colors represented as hex code
-class Hex(Color):
-    def __init__(self, hex_code):
-        if not hex_code[0] == '#':
-            hex_code = '#' + hex_code
-        self.hex_code = hex_code.upper()
-
-    def __str__(self):
-        return self.hex_code
-
-    def __sub__(self, other):
-        return super().__sub__(other)
-
-    def to_RGB(self):
-        return RGB(
-            int(self.hex_code[1:3], 16),
-            int(self.hex_code[3:5], 16),
-            int(self.hex_code[5:7], 16))
-
-    def to_Hex(self):
-        return self
-
-    def to_OKLCH(self):
-        return self.to_RGB().to_OKLCH()
-
-    # Check whether the color is in-gamut
-    def is_in_gamut(self):
-        return super().is_in_gamut()
-
-# OKLCH colors represented as triplets
-class OKLCH(Color):
-    def __init__(self, l, c, h):
-        self.l = l
-        self.c = c
-        self.h = h
-
-    def __str__(self):
-        return "oklch({}, {}, {})".format(self.l, self.c, self.h)
-
-    def __sub__(self, other):
-        return super().__sub__(other)
 
     # Functions for converting to linear RGB from standard RGB and vice versa
     @staticmethod
@@ -371,53 +325,176 @@ class OKLCH(Color):
         else:
             return x / 12.92
 
-    # Convert from OKLAB
+    def to_OKLAB(self):
+        l = 0.4122214708 * self._f_inv(self.r/255) \
+                + 0.5363325363 * self._f_inv(self.g/255) \
+                + 0.0514459929 * self._f_inv(self.b/255)
+        m = 0.2119034982 * self._f_inv(self.r/255) \
+                + 0.6806995451 * self._f_inv(self.g/255) \
+                + 0.1073969566 * self._f_inv(self.b/255)
+        s = 0.0883024619 * self._f_inv(self.r/255) \
+                + 0.2817188376 * self._f_inv(self.g/255) \
+                + 0.6299787005 * self._f_inv(self.b/255)
+
+        l_ = math.pow(l, 1/3)
+        m_ = math.pow(m, 1/3)
+        s_ = math.pow(s, 1/3)
+
+        return OKLAB(
+            0.2104542553*l_ + 0.7936177850*m_ - 0.0040720468*s_,
+            1.9779984951*l_ - 2.4285922050*m_ + 0.4505937099*s_,
+            0.0259040371*l_ + 0.7827717662*m_ - 0.8086757660*s_)
+    def to_OKLCH(self):
+        return self.to_OKLAB().to_OKLCH()
+
+    # Check whether the color is in-gamut
+    def is_in_gamut(self):
+        return max(self.r, self.g, self.b) <= 255 and \
+                min(self.r, self.g, self.b) >= 0
+
+# RGB colors represented as hex code
+class HEX(Color):
+    def __init__(self, hex_code):
+        if not hex_code[0] == '#':
+            hex_code = '#' + hex_code
+        self.hex_code = hex_code.upper()
+
+    def __str__(self):
+        return self.hex_code
+
+    # Return type for addition and subtraction is type of first operand, unless
+    #   the result is an invalid color, as this could result in a mangled hex
+    #   code which is ambiguous. Therefore, these are left as RGB. 
+    # The below function sorts this out:
     @staticmethod
-    def _polarize(l, a, b):
-        c = math.pow(a*a + b*b, 1/2)
-        h = math.degrees(math.atan2(b, a))
-        if h < 0:
-            h += 360
+    def __get_valid_hex_or_rgb(color):
+        if color.is_in_gamut():
+            return color.to_HEX()
+        else: return color.to_RGB()
+    def __add__(self, other):
+        return self.__get_valid_hex_or_rgb(super().__add__(other))
+    def __neg__(self):
+        return self.__get_valid_hex_or_rgb(super().__neg__())
+    def __sub__(self, other):
+        return self.__get_valid_hex_or_rgb(super().__sub__(other))
 
-        return OKLCH(l, c, h)
-    # Convert back to rectangular coords; this function leaves a&b normalized
-    @staticmethod
-    def _rect_normal(hue):
-        a = math.cos(math.radians(hue))
-        b = math.sin(math.radians(hue))
+    def __or__(self, other):
+        return super().__or__(other)
 
-        return a, b
-    # Multiply by c to get OKLAB components
-    def _rect(self):
-        a, b = self._rect_normal(self.h)
-
-        return a * self.c, b * self.c
-
+    # Type Conversions
     def to_RGB(self):
-        a, b = self._rect()
+        return RGB(
+            int(self.hex_code[1:3], 16),
+            int(self.hex_code[3:5], 16),
+            int(self.hex_code[5:7], 16))
+    def to_HEX(self):
+        return self
+    def to_OKLAB(self):
+        return self.to_OKLCH().to_OKLAB()
+    def to_OKLCH(self):
+        return self.to_RGB().to_OKLCH()
 
-        l_ = self.l + 0.3963377774 * a + 0.2158037573 * b
-        m_ = self.l - 0.1055613458 * a - 0.0638541728 * b
-        s_ = self.l - 0.0894841775 * a - 1.2914855480 * b
+    # Check whether the color is in-gamut
+    def is_in_gamut(self):
+        return super().is_in_gamut()
+
+# OKLAB colors represented as triplets
+class OKLAB(Color):
+    def __init__(self, l, a, b):
+        self.l = l
+        self.a = a
+        self.b = b
+
+    def __str__(self):
+        return f"oklab({self.l}, {self.a}, {self.b})"
+
+    # Return type for addition and subtraction is type of first operand
+    def __add__(self, other):
+        l = 0.5*(self.l + other.l)
+        a = 0.5*(self.a + other.a)
+        b = 0.5*(self.b + other.b)
+
+        return OKLAB(l, a, b)
+    def __neg__(self):
+        return OKLAB(1-self.l, -self.a, -self.b)
+    def __sub__(self, other):
+        return self + other.__neg__()
+
+    def __or__(self, other):
+        return super().__or__(other)
+
+    # Type Conversions
+    def to_RGB(self):
+        l_ = self.l + 0.3963377774 * self.a + 0.2158037573 * self.b
+        m_ = self.l - 0.1055613458 * self.a - 0.0638541728 * self.b
+        s_ = self.l - 0.0894841775 * self.a - 1.2914855480 * self.b
 
         l = l_*l_*l_
         m = m_*m_*m_
         s = s_*s_*s_
 
         return RGB(
-            _round(OKLCH._f(+4.0767416621 * l \
+            _round(RGB._f(+4.0767416621 * l \
                     - 3.3077115913 * m \
                     + 0.2309699292 * s) * 255),
-            _round(OKLCH._f(-1.2684380046 * l \
+            _round(RGB._f(-1.2684380046 * l \
                     + 2.6097574011 * m \
                     - 0.3413193965 * s) * 255),
-            _round(OKLCH._f(-0.0041960863 * l \
+            _round(RGB._f(-0.0041960863 * l \
                     - 0.7034186147 * m \
                     + 1.7076147010 * s) * 255))
+    def to_HEX(self):
+        return self.to_RGB().to_HEX()
+    def to_OKLAB(self):
+        return self
+    def to_OKLCH(self):
+        c = math.pow(self.a ** 2 + self.b ** 2, 0.5)
+        h = math.degrees(math.atan2(self.b, self.a))
+        if h < 0:
+            h += 360
 
-    def to_Hex(self):
-        return self.to_RGB().to_Hex()
+        return OKLCH(self.l, c, h)
 
+    def is_in_gamut(self):
+        return super().is_in_gamut()
+
+# OKLCH colors represented as triplets
+class OKLCH(Color):
+    def __init__(self, l, c, h):
+        self.l = l
+        self.c = c
+        self.h = h
+
+    def __str__(self):
+        return "oklch({}, {}, {})".format(self.l, self.c, self.h)
+
+    # Return type for addition and subtraction is type of first operand
+    def __add__(self, other):
+        return super().__add__(other).to_OKLCH()
+    def __neg__(self):
+        return super().__neg__().to_OKLCH()
+    def __sub__(self, other):
+        return super().__sub__(other).to_OKLCH()
+
+    def __or__(self, other):
+        return super().__or__(other)
+
+    # Type Conversions
+    def to_RGB(self):
+        return self.to_OKLAB().to_RGB()
+    def to_HEX(self):
+        return self.to_RGB().to_HEX()
+
+    @staticmethod
+    def _get_normalized_ab(hue):
+        a = math.cos(math.radians(hue))
+        b = math.sin(math.radians(hue))
+
+        return a, b
+    def to_OKLAB(self):
+        a, b = self._get_normalized_ab(self.h)
+
+        return OKLAB(self.l, a * self.c, b * self.c)
     def to_OKLCH(self):
         return self